+++ core-ng-search/src/main/java/core/framework/search/UpdateRequest.java
@@ -1,0 +1,11 @@
+package core.framework.search;
+
+/**
+ * @author neo
+ */
+public class UpdateRequest<T> {
+    public String index;
+    public String id;
+    public T source;
+    public String script;
+}