+++ core-ng-search/src/test/java/core/framework/search/impl/DocumentClassValidatorTest.java
@@ -1,6 +1,13 @@
 package core.framework.search.impl;
 
+import core.framework.api.json.Property;
+import core.framework.api.validate.NotNull;
+import core.framework.search.Index;
 import org.junit.jupiter.api.Test;
+
+import java.time.LocalDateTime;
+import java.util.List;
+import java.util.Map;
 
 /**
  * @author neo
@@ -9,5 +16,21 @@ class DocumentClassValidatorTest {
     @Test
     void validate() {
         new DocumentClassValidator(TestDocument.class).validate();
+    }
+
+    @Index(index = "main", type = "test")
+    public static class TestDocument {
+        @Property(name = "date_time_field")
+        public LocalDateTime dateTimeField;
+
+        @NotNull
+        @Property(name = "string_field")
+        public String stringField;
+
+        @Property(name = "list_field")
+        public List<String> listField;
+
+        @Property(name = "map_field")
+        public Map<String, String> mapField;
     }
 }