+++ CHANGELOG.md
@@ -1,5 +1,6 @@
 ## Change log
 ### 6.6.1 (8/5/2018 - )
+* search: support update by partial doc or script 
 
 ### 6.6.0 (7/31/2018 - 8/5/2018)   !!! check break changes !!!
 * db: !!! replaced repository.update with repository.partialUpdate(), repository.update will update all fields and field can be NULL